requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
discord.py==1.7.3
idna==3.3
multidict==6.0.2
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
typing-extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
